requirements.txt
jupyterlab==4.0.2
GitPython==3.1.31